Move LLM integration from bot/rosey to lib directory
- Move llm package from bot/rosey/llm to lib/llm
- Update imports in rosey.py: bot.rosey.llm -> lib.llm
- Update imports in test_llm.py
- Update imports in test_openai_provider.py

Rationale: LLM integration is core infrastructure that could be
used by other bots or components, not specific to rosey bot.
