+++ bot/rosey/test_llm.py
@@ -14,7 +14,7 @@ import json
 project_root = Path(__file__).parent.parent.parent
 sys.path.insert(0, str(project_root))
 
-from bot.rosey.llm import LLMClient
+from lib.llm import LLMClient
 
 
 async def test_ollama():
